Added PokemonRepository The repo now does all of the 'talking' between the Api and the Database
