+++ PokeDex/app/src/main/java/com/minneydev/pokedex/worker/DownloadPokemonWorker.kt
@@ -6,6 +6,7 @@ import androidx.work.Worker
 import androidx.work.WorkerParameters
 import com.minneydev.pokedex.App
 import com.minneydev.pokedex.MainActivity
+import com.minneydev.pokedex.repository.PokemonRepository
 import com.minneydev.pokedex.util.PokemonManager
 import com.minneydev.pokedex.util.PokemonManager.Companion.currentGen
 import kotlinx.coroutines.*
@@ -13,7 +14,7 @@ import kotlinx.coroutines.*
 class DownloadPokemonWorker(context: Context, workerParameters: WorkerParameters) :
                     Worker(context, workerParameters) {
 
-    private val pokemonManager by lazy { PokemonManager() }
+    private val pokemonRepository by lazy { PokemonRepository() }
     private val firstGenRange = 1..151
     private val secondGenRange = 152..251
     private val thirdGenRange = 252..386
@@ -37,8 +38,8 @@ class DownloadPokemonWorker(context: Context, workerParameters: WorkerParameters
         }
         CoroutineScope(Dispatchers.Main).launch {
             for (i in range) {
-                with (App.pokemonApi.fetchPokemonById("$i")) {
-                    pokemonManager.savePokemon(this)?.let { pokemon ->
+                with (pokemonRepository.fetchPokemonById(i.toString())) {
+                    pokemonRepository.savePokemon(this)?.let { pokemon ->
                         MainActivity.setPokemon(pokemon)
                     }
                     Log.d("SAVING", "$i: ${this?.name}")
