+++ PokeDex/app/src/main/java/com/minneydev/pokedex/worker/RefreshPokemonWorker.kt
@@ -5,35 +5,27 @@ import android.util.Log
 import android.widget.Toast
 import androidx.work.Worker
 import androidx.work.WorkerParameters
-import com.minneydev.pokedex.App
 import com.minneydev.pokedex.MainActivity
 import com.minneydev.pokedex.R
-import com.minneydev.pokedex.util.PokemonManager
+import com.minneydev.pokedex.repository.PokemonRepository
 import com.minneydev.pokedex.util.PokemonManager.Companion.currentGen
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 
 class RefreshPokemonWorker(context: Context, workerParameters: WorkerParameters) :
                            Worker(context, workerParameters) {
 
-    private val pokemonManager by lazy { PokemonManager() }
+    private val pokemonRepository by lazy { PokemonRepository() }
 
     override fun doWork(): Result {
         currentGen.let {
             Toast.makeText(applicationContext, R.string.refresh_toast, Toast.LENGTH_SHORT).show()
-            nukeDatabase()
+            pokemonRepository.nukeDatabase()
             reDownloadPokemon(it)
             Result.success()
         }
         return Result.failure()
-    }
-
-    private fun nukeDatabase() {
-        CoroutineScope(Dispatchers.IO).launch {
-            App.pokemonDb.pokemonDao().nukeTable()
-        }
     }
 
     private fun reDownloadPokemon(gen: Int) {
@@ -46,8 +38,8 @@ class RefreshPokemonWorker(context: Context, workerParameters: WorkerParameters)
         }
         CoroutineScope(Dispatchers.Main).launch {
             for (i in range) {
-                with (App.pokemonApi.fetchPokemonById("$i")) {
-                    pokemonManager.savePokemon(this)?.let { pokemon ->
+                with (pokemonRepository.fetchPokemonById(i.toString())) {
+                    pokemonRepository.savePokemon(this)?.let { pokemon ->
                         MainActivity.setPokemon(pokemon)
                     }
                     Log.d("PERIODIC", "$i: ${this?.name}")
